requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-google-genai
langchain-groq
chromadb
pypdf
pdfplumber
pytesseract
Pillow
sentence-transformers
torch
transformers
ollama
python-dotenv